chore: refactor ZeroMQ-related code to a separate file
